Rename xap_bsc_dev to match the class name.
