+++ lib/xap/schema/xap_bsc_device.rb
